// processors/kriya-clmm/src/types/sui/0xf6c05e2d9301e6e91dc6ab6c3ca918f7d55896e1f1edd64adc0e615cde27ebf1.ts
export class admin extends SuiBaseProcessor {
  constructor(options: SuiBindOptions) {
    super("admin", options);
  }
  static DEFAULT_OPTIONS: SuiBindOptions = {
    address:
      "0xf6c05e2d9301e6e91dc6ab6c3ca918f7d55896e1f1edd64adc0e615cde27ebf1",
    network: SuiNetwork.MAIN_NET,
  };

  static bind(options: Partial<SuiBindOptions> = {}): admin {
    return new admin({ ...admin.DEFAULT_OPTIONS, ...options });
  }

  onEventCollectProtocolFeeEvent(
    func: (
      event: admin.CollectProtocolFeeEventInstance,
      ctx: SuiContext,
    ) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): admin {
    this.onMoveEvent(
      func,
      { type: "admin::CollectProtocolFeeEvent" },
      fetchConfig,
    );
    return this;
  }

  onEventSetProtocolFlashLoanFeeRateEvent(
    func: (
      event: admin.SetProtocolFlashLoanFeeRateEventInstance,
      ctx: SuiContext,
    ) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): admin {
    this.onMoveEvent(
      func,
      { type: "admin::SetProtocolFlashLoanFeeRateEvent" },
      fetchConfig,
    );
    return this;
  }

  onEventSetProtocolSwapFeeRateEvent(
    func: (
      event: admin.SetProtocolSwapFeeRateEventInstance,
      ctx: SuiContext,
    ) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): admin {
    this.onMoveEvent(
      func,
      { type: "admin::SetProtocolSwapFeeRateEvent" },
      fetchConfig,
    );
    return this;
  }
}

export namespace admin {
  export interface CollectProtocolFeeEvent {
    sender: string;
    pool_id: _0x2.object$.ID;
    amount_x: bigint;
    amount_y: bigint;
  }

  export namespace CollectProtocolFeeEvent {
    export const TYPE_QNAME =
      "0xf6c05e2d9301e6e91dc6ab6c3ca918f7d55896e1f1edd64adc0e615cde27ebf1::admin::CollectProtocolFeeEvent";

    const TYPE = new TypeDescriptor<CollectProtocolFeeEvent>(
      CollectProtocolFeeEvent.TYPE_QNAME,
    );

    export function type(): TypeDescriptor<CollectProtocolFeeEvent> {
      return TYPE.apply();
    }
  }

  export type CollectProtocolFeeEventInstance =
    TypedEventInstance<CollectProtocolFeeEvent> & {
      data_decoded: CollectProtocolFeeEvent;
      type_arguments: [];
    };

  export interface SetProtocolFlashLoanFeeRateEvent {
    sender: string;
    pool_id: _0x2.object$.ID;
    protocol_fee_share_old: bigint;
    protocol_fee_share_new: bigint;
  }
// ...
  export type SetProtocolFlashLoanFeeRateEventInstance =
    TypedEventInstance<SetProtocolFlashLoanFeeRateEvent> & {
      data_decoded: SetProtocolFlashLoanFeeRateEvent;
      type_arguments: [];
    };
// ...
}
// ...
export class create_pool extends SuiBaseProcessor {
  constructor(options: SuiBindOptions) {
    super("create_pool", options);
  }
  static DEFAULT_OPTIONS: SuiBindOptions = {
    address:
      "0xf6c05e2d9301e6e91dc6ab6c3ca918f7d55896e1f1edd64adc0e615cde27ebf1",
    network: SuiNetwork.MAIN_NET,
  };

  static bind(options: Partial<SuiBindOptions> = {}): create_pool {
    return new create_pool({ ...create_pool.DEFAULT_OPTIONS, ...options });
  }

  onEventPoolCreatedEvent(
    func: (
      event: create_pool.PoolCreatedEventInstance,
      ctx: SuiContext,
    ) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): create_pool {
    this.onMoveEvent(
      func,
      { type: "create_pool::PoolCreatedEvent" },
      fetchConfig,
    );
    return this;
  }
}

export namespace create_pool {
  export interface PoolCreatedEvent {
    sender: string;
    pool_id: _0x2.object$.ID;
    type_x: _0x1.type_name.TypeName;
    type_y: _0x1.type_name.TypeName;
    fee_rate: bigint;
    tick_spacing: number;
  }

  export namespace PoolCreatedEvent {
    export const TYPE_QNAME =
      "0xf6c05e2d9301e6e91dc6ab6c3ca918f7d55896e1f1edd64adc0e615cde27ebf1::create_pool::PoolCreatedEvent";

    const TYPE = new TypeDescriptor<PoolCreatedEvent>(
      PoolCreatedEvent.TYPE_QNAME,
    );

    export function type(): TypeDescriptor<PoolCreatedEvent> {
      return TYPE.apply();
    }
  }

  export type PoolCreatedEventInstance =
    TypedEventInstance<PoolCreatedEvent> & {
      data_decoded: PoolCreatedEvent;
      type_arguments: [];
    };

  export namespace builder {
    export function new$<T0 = any, T1 = any>(
      tx: Transaction,
      args: [
        string | TransactionObjectArgument | TransactionArgument,
        bigint | TransactionArgument,
        string | TransactionObjectArgument | TransactionArgument,
      ],
      typeArguments: [TypeDescriptor<T0> | string, TypeDescriptor<T1> | string],
    ): TransactionArgument &
      [TransactionArgument, TransactionArgument, TransactionArgument] {
      const _args: any[] = [];
      _args.push(transactionArgumentOrObject(args[0], tx));
      _args.push(transactionArgumentOrPureU64(args[1], tx));
      _args.push(transactionArgumentOrObject(args[2], tx));

      // @ts-ignore
      return tx.moveCall({
        target:
          "0xf6c05e2d9301e6e91dc6ab6c3ca918f7d55896e1f1edd64adc0e615cde27ebf1::create_pool::new",
        arguments: _args,
        typeArguments: [
          typeof typeArguments[0] === "string"
            ? typeArguments[0]
            : typeArguments[0].getSignature(),
          typeof typeArguments[1] === "string"
            ? typeArguments[1]
            : typeArguments[1].getSignature(),
        ],
      });
    }
  }
  export namespace view {
    export async function new$<T0 = any, T1 = any>(
      client: SuiClient,
      args: [string, bigint, string],
      typeArguments: [TypeDescriptor<T0> | string, TypeDescriptor<T1> | string],
    ): Promise<TypedDevInspectResults<[pool.Pool<T0, T1>]>> {
      const tx = new Transaction();
      builder.new$(tx, args, typeArguments);
      const inspectRes = await client.devInspectTransactionBlock({
        transactionBlock: tx,
        sender: ZERO_ADDRESS,
      });

      return (await getMoveCoder(client)).decodeDevInspectResult<
        [pool.Pool<T0, T1>]
      >(inspectRes);
    }
  }
}
// ...
export class liquidity extends SuiBaseProcessor {
  constructor(options: SuiBindOptions) {
    super("liquidity", options);
  }
  static DEFAULT_OPTIONS: SuiBindOptions = {
    address:
      "0xf6c05e2d9301e6e91dc6ab6c3ca918f7d55896e1f1edd64adc0e615cde27ebf1",
    network: SuiNetwork.MAIN_NET,
  };

  static bind(options: Partial<SuiBindOptions> = {}): liquidity {
    return new liquidity({ ...liquidity.DEFAULT_OPTIONS, ...options });
  }

  onEventAddLiquidityEvent(
    func: (event: liquidity.AddLiquidityEventInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): liquidity {
    this.onMoveEvent(
      func,
      { type: "liquidity::AddLiquidityEvent" },
      fetchConfig,
    );
    return this;
  }

  onEventClosePositionEvent(
    func: (
      event: liquidity.ClosePositionEventInstance,
      ctx: SuiContext,
    ) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): liquidity {
    this.onMoveEvent(
      func,
      { type: "liquidity::ClosePositionEvent" },
      fetchConfig,
    );
    return this;
  }

  onEventOpenPositionEvent(
    func: (event: liquidity.OpenPositionEventInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): liquidity {
    this.onMoveEvent(
      func,
      { type: "liquidity::OpenPositionEvent" },
      fetchConfig,
    );
    return this;
  }

  onEventRemoveLiquidityEvent(
    func: (
      event: liquidity.RemoveLiquidityEventInstance,
      ctx: SuiContext,
    ) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): liquidity {
    this.onMoveEvent(
      func,
      { type: "liquidity::RemoveLiquidityEvent" },
      fetchConfig,
    );
    return this;
  }
}
// ...
export class pool extends SuiBaseProcessor {
  constructor(options: SuiBindOptions) {
    super("pool", options);
  }
  static DEFAULT_OPTIONS: SuiBindOptions = {
    address:
      "0xf6c05e2d9301e6e91dc6ab6c3ca918f7d55896e1f1edd64adc0e615cde27ebf1",
    network: SuiNetwork.MAIN_NET,
  };

  static bind(options: Partial<SuiBindOptions> = {}): pool {
    return new pool({ ...pool.DEFAULT_OPTIONS, ...options });
  }

  onEventObservationCardinalityUpdatedEvent(
    func: (
      event: pool.ObservationCardinalityUpdatedEventInstance,
      ctx: SuiContext,
    ) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): pool {
    this.onMoveEvent(
      func,
      { type: "pool::ObservationCardinalityUpdatedEvent" },
      fetchConfig,
    );
    return this;
  }

  onEventPoolRewardCustodianDfKey(
    func: (
      event: pool.PoolRewardCustodianDfKeyInstance,
      ctx: SuiContext,
    ) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): pool {
    this.onMoveEvent(
      func,
      { type: "pool::PoolRewardCustodianDfKey" },
      fetchConfig,
    );
    return this;
  }

  onEventPoolRewardInfo(
    func: (event: pool.PoolRewardInfoInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): pool {
    this.onMoveEvent(func, { type: "pool::PoolRewardInfo" }, fetchConfig);
    return this;
  }

  onEventUpdatePoolRewardEmissionEvent(
    func: (
      event: pool.UpdatePoolRewardEmissionEventInstance,
      ctx: SuiContext,
    ) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): pool {
    this.onMoveEvent(
      func,
      { type: "pool::UpdatePoolRewardEmissionEvent" },
      fetchConfig,
    );
    return this;
  }
}
// ...
export class tick extends SuiBaseProcessor {
  constructor(options: SuiBindOptions) {
    super("tick", options);
  }
  static DEFAULT_OPTIONS: SuiBindOptions = {
    address:
      "0xf6c05e2d9301e6e91dc6ab6c3ca918f7d55896e1f1edd64adc0e615cde27ebf1",
    network: SuiNetwork.MAIN_NET,
  };

  static bind(options: Partial<SuiBindOptions> = {}): tick {
    return new tick({ ...tick.DEFAULT_OPTIONS, ...options });
  }

  onEventTickInfo(
    func: (event: tick.TickInfoInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): tick {
    this.onMoveEvent(func, { type: "tick::TickInfo" }, fetchConfig);
    return this;
  }
}
// ...
export class trade extends SuiBaseProcessor {
  constructor(options: SuiBindOptions) {
    super("trade", options);
  }
  static DEFAULT_OPTIONS: SuiBindOptions = {
    address:
      "0xf6c05e2d9301e6e91dc6ab6c3ca918f7d55896e1f1edd64adc0e615cde27ebf1",
    network: SuiNetwork.MAIN_NET,
  };

  static bind(options: Partial<SuiBindOptions> = {}): trade {
    return new trade({ ...trade.DEFAULT_OPTIONS, ...options });
  }

  onEventFlashLoanEvent(
    func: (event: trade.FlashLoanEventInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): trade {
    this.onMoveEvent(func, { type: "trade::FlashLoanEvent" }, fetchConfig);
    return this;
  }

  onEventRepayFlashLoanEvent(
    func: (event: trade.RepayFlashLoanEventInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): trade {
    this.onMoveEvent(func, { type: "trade::RepayFlashLoanEvent" }, fetchConfig);
    return this;
  }

  onEventRepayFlashSwapEvent(
    func: (event: trade.RepayFlashSwapEventInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): trade {
    this.onMoveEvent(func, { type: "trade::RepayFlashSwapEvent" }, fetchConfig);
    return this;
  }

  onEventSwapEvent(
    func: (event: trade.SwapEventInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): trade {
    this.onMoveEvent(func, { type: "trade::SwapEvent" }, fetchConfig);
    return this;
  }

  onEventSwapState(
    func: (event: trade.SwapStateInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): trade {
    this.onMoveEvent(func, { type: "trade::SwapState" }, fetchConfig);
    return this;
  }

  onEventSwapStepComputations(
    func: (event: trade.SwapStepComputationsInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): trade {
    this.onMoveEvent(
      func,
      { type: "trade::SwapStepComputations" },
      fetchConfig,
    );
    return this;
  }
}